Move all RpcValue conversions to RpcValueConverter
