+++ src/WinUIShell.Common/RpcService.cs
@@ -133,7 +133,7 @@ internal sealed class RpcService
                     var obj = ObjectStore.Get().GetObject(id);
                     var arguments = ConvertArguments(rpcArguments);
                     object? result = Invoker.Get().InvokeMethod(obj, methodName, arguments);
-                    taskCompletion.SetResult(ConvertObjectToRpcValue(result));
+                    taskCompletion.SetResult(RpcValueConverter.ConvertObjectToRpcValue(result));
                 }
                 catch (Exception e)
                 {
@@ -188,7 +188,7 @@ internal sealed class RpcService
                 {
                     var arguments = ConvertArguments(rpcArguments);
                     object? result = Invoker.Get().InvokeStaticMethod(className, methodName, arguments);
-                    taskCompletion.SetResult(ConvertObjectToRpcValue(result));
+                    taskCompletion.SetResult(RpcValueConverter.ConvertObjectToRpcValue(result));
                 }
                 catch (Exception e)
                 {
@@ -208,7 +208,7 @@ internal sealed class RpcService
         object?[] arguments = new object[rpcArguments.Length];
         for (int i = 0; i < arguments.Length; ++i)
         {
-            arguments[i] = ConvertRpcValueToObject(rpcArguments[i]);
+            arguments[i] = RpcValueConverter.ConvertRpcValueToObject(rpcArguments[i]);
         }
         return arguments;
     }
@@ -220,7 +220,7 @@ internal sealed class RpcService
             () =>
             {
                 var obj = ObjectStore.Get().GetObject(id);
-                var value = ConvertRpcValueToObject(rpcValue);
+                var value = RpcValueConverter.ConvertRpcValueToObject(rpcValue);
                 Invoker.Get().SetProperty(obj, propertyName, value);
             });
     }
@@ -235,7 +235,7 @@ internal sealed class RpcService
                 try
                 {
                     var obj = ObjectStore.Get().GetObject(id);
-                    var value = ConvertRpcValueToObject(rpcValue);
+                    var value = RpcValueConverter.ConvertRpcValueToObject(rpcValue);
                     Invoker.Get().SetProperty(obj, propertyName, value);
 
                     taskCompletion.SetResult();
@@ -259,7 +259,7 @@ internal sealed class RpcService
                 {
                     var obj = ObjectStore.Get().GetObject(id);
                     var value = Invoker.Get().GetProperty(obj, propertyName);
-                    taskCompletion.SetResult(ConvertObjectToRpcValue(value));
+                    taskCompletion.SetResult(RpcValueConverter.ConvertObjectToRpcValue(value));
                 }
                 catch (Exception e)
                 {
@@ -279,7 +279,7 @@ internal sealed class RpcService
                 try
                 {
                     var value = Invoker.Get().GetStaticProperty(className, propertyName);
-                    taskCompletion.SetResult(ConvertObjectToRpcValue(value));
+                    taskCompletion.SetResult(RpcValueConverter.ConvertObjectToRpcValue(value));
                 }
                 catch (Exception e)
                 {
@@ -296,8 +296,8 @@ internal sealed class RpcService
             () =>
             {
                 var obj = ObjectStore.Get().GetObject(id);
-                var index = ConvertRpcValueToObject(rpcIndex);
-                var value = ConvertRpcValueToObject(rpcValue);
+                var index = RpcValueConverter.ConvertRpcValueToObject(rpcIndex);
+                var value = RpcValueConverter.ConvertRpcValueToObject(rpcValue);
 
                 if (index is null)
                 {
@@ -318,14 +318,14 @@ internal sealed class RpcService
                 try
                 {
                     var obj = ObjectStore.Get().GetObject(id);
-                    var index = ConvertRpcValueToObject(rpcIndex);
+                    var index = RpcValueConverter.ConvertRpcValueToObject(rpcIndex);
                     if (index is null)
                     {
                         throw new InvalidOperationException("Index of Indexer property cannot be null.");
                     }
 
                     var value = Invoker.Get().GetIndexerProperty(obj, index);
-                    taskCompletion.SetResult(ConvertObjectToRpcValue(value));
+                    taskCompletion.SetResult(RpcValueConverter.ConvertObjectToRpcValue(value));
                 }
                 catch (Exception e)
                 {
@@ -339,86 +339,5 @@ internal sealed class RpcService
     {
         // Do not use CommandServer to show errors immediately without being blocked by other commands.
         Console.Error.WriteLine(message);
-    }
-
-    private object? ConvertRpcValueToEnum(RpcValue rpcValue)
-    {
-        var value = rpcValue.GetObject();
-        if (value is null)
-        {
-            return null;
-        }
-
-        var sourceEnumName = rpcValue.GetEnumTypeName();
-        if (sourceEnumName is null)
-        {
-            return null;
-        }
-
-        _ = EnumTypeMapping.Get().TryGetValue(sourceEnumName, out string? enumTargetName);
-        if (enumTargetName is null)
-        {
-            throw new InvalidOperationException($"Enum mapping for [{sourceEnumName}] not found.");
-        }
-
-        var targetEnumType = Type.GetType(enumTargetName);
-        if (targetEnumType == null)
-        {
-            throw new InvalidOperationException($"Type [{enumTargetName}] not found.");
-        }
-
-        return Enum.ToObject(targetEnumType, value);
-    }
-
-    private object? ConvertRpcValueToObject(RpcValue rpcValue)
-    {
-        var enumValue = ConvertRpcValueToEnum(rpcValue);
-        if (enumValue is not null)
-        {
-            return enumValue;
-        }
-
-        var value = rpcValue.GetObject();
-        if (value is RpcValue[] array)
-        {
-            var objectArray = new object?[array.Length];
-            for (int i = 0; i < array.Length; ++i)
-            {
-                objectArray[i] = ConvertRpcValueToObject(array[i]);
-            }
-            return objectArray;
-        }
-        else
-        if (value is ObjectId objectId)
-        {
-            return ObjectStore.Get().GetObject(objectId);
-        }
-        else
-        {
-            return value;
-        }
-    }
-
-    private RpcValue ConvertObjectToRpcValue(object? obj)
-    {
-        if (RpcValue.IsSupportedType(obj))
-        {
-            return new RpcValue(obj);
-        }
-        else
-        {
-            var valueObjectId = ObjectStore.Get().FindId(obj!);
-            if (valueObjectId is not null)
-            {
-                return new RpcValue(valueObjectId);
-            }
-            else
-            {
-                // If the object is not a primitive type or a registered object, register it here.
-                // The corresponding object needs to be created on the client side.
-                _ = ObjectStore.Get().RegisterObjectWithType(obj!, out ObjectId id);
-                return new RpcValue(id);
-            }
-        }
     }
 }
